+++ src/app/cart/models/cart-item.model.ts
@@ -2,6 +2,7 @@ import { ProductModel } from '../../products/models/product.model';
 
 export interface CartModel {
   item: ProductModel;
+  name: string;
   count: number;
   price: number;
 }